WIP: hardcode the replacement knot selector to `splrep`, check vs fitpack
The test failure in test_fitpack/check_1 for k=2 is curious: it is due
to the sin function being symmetric. With adding `nplus=2` knots,
at the 2nd step the sum of residuals is symmetric for knot intervals
[pi/2, pi] and [pi, 3*pi/2]. FITPACK and make_lsq_spline give the sums
of residuals equal up to numerical noise (which is OK) --- and then
this numerical noise decides where to place the next knot.

The net effect is that the knot vectors for  `construct_knot_vector`
and fpcurf.f are different.