ENH: interpolate: add make_splPrep for parametric spline fitting
Reuse make_splrep, plumb parametric=True/False parameter.
The main (the only) difference is in fact the way the residuals are
computed: in the D > 1 parametric case, deviations are summed over
the dimensions.

To simplify implementations, make _impl functions always see and work
with y.ndim == 2; squeeze the extra dim out in postprocessing if needed.